+++ src/controllers/mer_controller.py
@@ -1,17 +1,14 @@
 import sys
-import threading
-from typing import List, Generator, Dict
+from typing import List, Dict
 
 from PyQt5.QtCore import QObject
 from PyQt5.QtWidgets import QApplication
-from pandas import DataFrame
-import pandas as pd
 
-from src.importers.bulk_importer import BulkImporter
-from src.importers.mer_importer import MerImporter
+from src.handlers.bulk_handler import BulkHandler
+from src.handlers.file_handler import FileHandler
 from src.models.dataframe_model import DataFrameModel
 from src.models.mer_model import MerModel
-from src.utility.utility import get_exception, get_files_from_folder
+from src.utility.utility import get_files_from_folder
 from src.views.bulk_export_dlg import BulkSettings
 from src.views.mer_view import MerView
 
@@ -23,8 +20,8 @@ class MerController(QObject):
         self.model: MerModel = MerModel()
         self.view: MerView = MerView()
 
-        self.bulk_importer: BulkImporter = BulkImporter()
-        self.mer_importer: MerImporter = MerImporter(parent=self)
+        self.bulk_handler: BulkHandler = BulkHandler()
+        self.file_handler: FileHandler = FileHandler(parent=self)
 
         self.init()
 
@@ -36,21 +33,33 @@ class MerController(QObject):
         self.view.tree.selection_changed_signal.connect(self.select_df)
         self.view.exit_signal.connect(self.exit_program)
 
-        self.bulk_importer.import_busy.connect(self.view.import_busy)
-        self.bulk_importer.import_failed.connect(self.on_task_failed)
+        self.bulk_handler.task_busy.connect(self.view.import_busy)
+        self.bulk_handler.task_failed.connect(self.on_task_failed)
 
-        self.mer_importer.task_busy.connect(self.view.import_busy)
-        self.mer_importer.task_failed.connect(self.on_task_failed)
-        self.mer_importer.task_finished.connect(self.on_task_success)
+        self.file_handler.task_busy.connect(self.view.import_busy)
+        self.file_handler.task_finished.connect(self.on_task_success)
+        self.file_handler.task_failed.connect(self.on_task_failed)
 
     def import_file(self, paths: List[str]) -> None:
         self.reset_mer()
 
         self.view.toggle_progress(True)
-        self.mer_importer.import_and_convert(paths)
+        self.file_handler.import_and_convert(paths)
 
     def import_bulk(self, info: BulkSettings):
-        self.bulk_importer.import_and_convert(list(get_files_from_folder(info.src)), info)
+        self.bulk_handler.import_and_convert(list(get_files_from_folder(info.src)), info)
+
+    def export(self, path: str) -> None:
+        selected_items: List[str] = list(self.view.tree.selected_items())
+        if len(selected_items) > 0:
+            data = dict()
+            for name in selected_items:
+                dfm = self.model.get_df(name)
+                data[name] = dfm
+
+            self.file_handler.start_export(data, path)
+        else:
+            self.view.import_failed('Select at least 1 Identifier')
 
     def on_task_failed(self, txt) -> None:
         self.reset_mer()
@@ -60,10 +69,7 @@ class MerController(QObject):
 
     def on_task_success(self, converted_data: Dict[str, DataFrameModel]) -> None:
         self.model.mer_data = converted_data
-        try:
-            self.set_mer_view(converted_data)
-        except Exception as e:
-            print('MerController.on_convert_success: ' + get_exception(e))
+        self.set_mer_view(converted_data)
 
     def set_mer_view(self, converted_data: Dict[str, DataFrameModel]) -> None:
         for name, idf in self.model.mer_data.items():
@@ -80,22 +86,6 @@ class MerController(QObject):
                 )
 
         self.view.import_success(tact_scenario_txt)
-
-    def export(self, path: str) -> None:
-        selected_items: Generator = self.view.tree.selected_items()
-        if len(list(selected_items)) > 0:
-            threading.Thread(target=self.export_to_xls, args=(path, self.view.tree.selected_items())).start()
-        else:
-            self.view.import_failed('Select at least 1 Identifier')
-
-    def export_to_xls(self, path: str, selected_items: Generator) -> None:
-        writer: pd.ExcelWriter = pd.ExcelWriter(path)
-
-        for name in selected_items:
-            df: DataFrame = self.model.get_df(name).df
-            df.to_excel(writer, name)
-
-        writer.save()
 
     def reset_mer(self) -> None:
         if self.model.has_mer():
